+++ ng-skos.js
@@ -115,7 +115,6 @@ angular.module('ngSKOS')
  * <li>prefLabel (object of strings)
  * <li>altLabel (object of array of strings)
  * <li>notation (string)
- * <li>note (object of array of strings)
  * <li>broader (array of concepts)
  * <li>narrower (array of concepts)
  * <li>related (array of concepts)
@@ -231,7 +230,7 @@ angular.module('ngSKOS')
         scope: { 
             label: '=skosLabel',
         },
-        template: '{{label[language] ? label[language] : label}}',
+        template: '{{label[language]}}',
         link: function(scope, element, attrs) {
 
             function updateLanguage(language) {
@@ -248,7 +247,7 @@ angular.module('ngSKOS')
 
             function selectLanguage(labels, language) {
                 if ( angular.isObject(labels) ) {
-                    if ( language && labels[language] ) {
+                    if ( language && angular.isString(labels[language])) {
                         return language;
                     } else {
                         return guessLanguage(labels);
@@ -259,7 +258,9 @@ angular.module('ngSKOS')
             function guessLanguage(labels) {
                 // TODO: https://github.com/gbv/ng-skos/issues/17
                 for (var language in labels) {
-                    return language; // take arbitrary language
+                    if (angular.isString(labels[language])) {
+                        return language; // take arbitrary language
+                    }
                 }
             }
 
@@ -267,9 +268,7 @@ angular.module('ngSKOS')
             attrs.$observe('lang', updateLanguage);
 
             // update if labels changed
-            scope.$watch('label', function(value) {
-                updateLanguage();
-            }, true);
+            scope.$watch('label', function() { updateLanguage(); }, true);
         },
     };
 });
@@ -380,6 +379,79 @@ angular.module('ngSKOS')
         }
     };
 }]);
+
+/**
+ * @ngdoc directive
+ * @name ng-skos.directive:skosNotes
+ * @restrict A
+ * @description
+ *
+ * Shows the [documentary notes](http://www.w3.org/TR/skos-primer/#secdocumentation)
+ * of a concept. A preferred language can be selected with parameter `lang`. If
+ * no language is selected, an arbitrary language with given notes is chosen.
+ *
+ * ## Scope
+ *
+ * The variables `scopeNote`, `definition`, `example`, and `historyNote` are
+ * added to the scope, if given es keys of the `skos-notes` parameter object.
+ *
+ * ## Customization
+ *
+ * The [default
+ * template](https://github.com/gbv/ng-skos/blob/master/src/templates/skos-notes.html) 
+ * can be changed with parameter `templateUrl`.
+ *
+ * @param {string} skos-notes Expression with multilingual notes data
+ * @param {string=} lang preferred language to show notes in.
+ * @param {string} template-url URL of a template to display the concept
+ */
+angular.module('ngSKOS')
+.directive('skosNotes', function() {
+    return {
+        restrict: 'A',
+        scope: { 
+            notes: '=skosNotes',
+        },
+        templateUrl: function(elem, attrs) {
+            return attrs.templateUrl ? 
+                   attrs.templateUrl : 'template/skos-notes.html';
+        },
+        link: function(scope, element, attrs) {
+
+            function update(language) {
+                scope.language = language ? language : attrs.lang;                
+                if (!language) { 
+                    language = guessLanguage(scope.notes);
+                }
+                if (language != scope.language) {
+                    scope.language = language;
+                }
+            }
+
+            // take arbitrary language
+            function guessLanguage(notes) {
+                for (var type in notes) {
+                    for (var language in notes[type]) {
+                        return language;
+                    }
+                }
+            }
+    
+            // update if lang attribute changed (also called at initialization)
+            attrs.$observe('lang', update);
+
+            // update if notes changed
+            scope.$watch('notes',function(notes) {
+                angular.forEach(['scopeNote','definition','example','historyNote'],
+                    function(field) {
+                        scope[field] = angular.isObject(notes) 
+                                     ? notes[field] : null;
+                    }
+                );
+            },true);
+        }
+    };
+});
 
 /**
  * @ngdoc directive
@@ -607,17 +679,22 @@ angular.module('ngSKOS').run(['$templateCache', function($templateCache) {
 
 
   $templateCache.put('template/skos-concept-thesaurus.html',
-    "<div class=\"skos-concept-thesaurus\"><ul ng-if=\"ancestors.length\" class=\"ancestors\"><li class=\"ancestor\" ng-repeat=\"a in ancestors\"><span class=\"ancestors-label\" skos-label=\"a.prefLabel\" lang=\"{{language}}\" ng-click=\"update(a);reload();\"></span> <span class=\"ancestors-greater\">></span></li></ul><div class=\"top top-classic\"><span ng-if=\"notation\" class=\"notation\">{{notation[0]}}</span> <b><span skos-label=\"concept.prefLabel\" lang=\"{{language}}\"></span></b><a ng-if=\"notation\" class=\"uri\" href=\"{{uri}}\"><span style=\"vertical-align:-10%\" class=\"glyphicon glyphicon-globe\"></span></a></div><div ng-if=\"altLabel\" class=\"skos-concept-altlabel\"><ul><li ng-repeat=\"a in altLabel[language]\" style=\"display:inline\"><span ng-if=\"$index < 5\" style=\"display:inline\">{{a}}</span> <span style=\"margin-left:-4px;margin-right:3px\" ng-if=\"$index < 4 && $index < altLabel[language].length-1\">,</span></li></ul></div><div ng-if=\"broader.length\" class=\"skos-concept-thesaurus-relation\"><b>Broader Terms:</b><ul ng-repeat=\"b in broader\"><li><span skos-label=\"b.prefLabel\" lang=\"{{language}}\" ng-click=\"click(b)\"></li></ul></div><div ng-if=\"narrower.length\" class=\"skos-concept-thesaurus-relation\"><b>Narrower Terms:</b><ul ng-repeat=\"n in narrower\"><li><span skos-label=\"n.prefLabel\" lang=\"{{language}}\" ng-click=\"click(n)\"></li></ul></div><div ng-if=\"related.length\" class=\"skos-concept-thesaurus-relation\"><b>Related Terms:</b><ul ng-repeat=\"r in related\"><li><span skos-label=\"r.prefLabel\" lang=\"{{language}}\" ng-click=\"click(r)\"></li></ul></div><div ng-if=\"scopeNote.length\"><div ng-repeat=\"n in scopeNote\" class=\"skos-concept-note skos-concept-thesaurus-note\"><em><div skos-label=\"n\" lang=\"{{language}}\"></div></em></div></div><div ng-if=\"definition.length\"><div ng-repeat=\"n in definition\" class=\"skos-concept-note skos-concept-thesaurus-note\"><em><div skos-label=\"n\" lang=\"{{language}}\"></div></em></div></div></div>"
+    "<div class=\"skos-concept-thesaurus\"><ul ng-if=\"ancestors.length\" class=\"ancestors\"><li class=\"ancestor\" ng-repeat=\"a in ancestors\"><span class=\"ancestors-label\" skos-label=\"a.prefLabel\" lang=\"{{language}}\" ng-click=\"update(a);reload();\"></span> <span class=\"ancestors-greater\">></span></li></ul><div class=\"top top-classic\"><span ng-if=\"notation\" class=\"notation\">{{notation[0]}}</span> <b><span skos-label=\"concept.prefLabel\" lang=\"{{language}}\"></span></b><a ng-if=\"notation\" class=\"uri\" href=\"{{uri}}\"><span style=\"vertical-align:-10%\" class=\"glyphicon glyphicon-globe\"></span></a></div><div ng-if=\"altLabel\" class=\"skos-concept-altlabel\"><ul><li ng-repeat=\"a in altLabel[language]\" style=\"display:inline\"><span ng-if=\"$index < 5\" style=\"display:inline\">{{a}}</span> <span style=\"margin-left:-4px;margin-right:3px\" ng-if=\"$index < 4 && $index < altLabel[language].length-1\">,</span></li></ul></div><div ng-if=\"broader.length\" class=\"skos-concept-thesaurus-relation\"><b>Broader Terms:</b><ul ng-repeat=\"b in broader\"><li><span skos-label=\"b.prefLabel\" lang=\"{{language}}\" ng-click=\"click(b)\"></li></ul></div><div ng-if=\"narrower.length\" class=\"skos-concept-thesaurus-relation\"><b>Narrower Terms:</b><ul ng-repeat=\"n in narrower\"><li><span skos-label=\"n.prefLabel\" lang=\"{{language}}\" ng-click=\"click(n)\"></li></ul></div><div ng-if=\"related.length\" class=\"skos-concept-thesaurus-relation\"><b>Related Terms:</b><ul ng-repeat=\"r in related\"><li><span skos-label=\"r.prefLabel\" lang=\"{{language}}\" ng-click=\"click(r)\"></li></ul></div><div ng-if=\"scopeNote.length\"><div ng-repeat=\"n in scopeNote\" class=\"skos-note skos-concept-thesaurus-note\"><em><div skos-label=\"n\" lang=\"{{language}}\"></div></em></div></div><div ng-if=\"definition.length\"><div ng-repeat=\"n in definition\" class=\"skos-note skos-concept-thesaurus-note\"><em><div skos-label=\"n\" lang=\"{{language}}\"></div></em></div></div></div>"
   );
 
 
   $templateCache.put('template/skos-concept.html',
-    "<div class=\"skos-concept\"><div class=\"top top-alt\"><span ng-if=\"notation.length\" class=\"notation\">{{notation[0]}}</span> <b><span ng-if=\"prefLabel\" skos-label=\"concept.prefLabel\" lang=\"{{language}}\"></span></b> <a ng-if=\"uri && uri != notation\" class=\"uri\" href=\"{{uri}}\" target=\"_blank\"><span class=\"glyphicon glyphicon-globe\"></span></a></div><div ng-if=\"altLabel\" class=\"skos-concept-altlabel\"><ul><li ng-repeat=\"a in altLabel[language]\"><span ng-if=\"$index < 5\">{{a}}</span> <span style=\"margin-left:-4px;margin-right:3px\" ng-if=\"$index < 4 && $index < altLabel[language].length-1\">,</span></li></ul></div><div ng-if=\"broader.length || narrower.length || related.length\" class=\"skos-concept-connected\"><div ng-if=\"broader.length\" class=\"skos-concept-relation skos-concept-relation-broader\"><ul ng-repeat=\"c in broader\"><li><div>&#8598;</div><div class=\"skos-concept-relation-label\"><span ng-if=\"c.prefLabel\" skos-label=\"c.prefLabel\" lang=\"{{language}}\" ng-click=\"click(c)\" title=\"{{c.notation[0]}}\"></span> <span ng-if=\"!c.prefLabel\">{{c.notation[0]}}</span></div></li></ul></div><div ng-if=\"narrower.length\" class=\"skos-concept-relation skos-concept-relation-narrower\"><ul ng-repeat=\"c in narrower\"><li><div>&#8600;</div><div class=\"skos-concept-relation-label\"><span ng-if=\"c.prefLabel\" skos-label=\"c.prefLabel\" lang=\"{{language}}\" ng-click=\"click(c)\" title=\"{{c.notation[0]}}\"></span> <span ng-if=\"!c.prefLabel\">{{c.notation[0]}}</span></div></li></ul></div><div ng-if=\"related.length\" class=\"skos-concept-relation skos-concept-relation-related\"><ul ng-repeat=\"c in related\"><li><div>&#8594;</div><div class=\"skos-concept-relation-label\"><span ng-if=\"c.prefLabel\" skos-label=\"c.prefLabel\" lang=\"{{language}}\" ng-click=\"click(c)\" title=\"{{c.notation[0]}}\"></span> <span ng-if=\"!c.prefLabel\">{{c.notation[0]}}</span></div></li></ul></div></div><div style=\"margin-bottom:10px\"><div ng-if=\"scopeNote.length\"><div ng-repeat=\"n in scopeNote\" class=\"skos-concept-note\"><em><div skos-label=\"n\" lang=\"{{language}}\"></div></em></div></div><div ng-if=\"definition.length\"><div ng-repeat=\"n in definition\" class=\"skos-concept-note\"><em><div skos-label=\"n\" lang=\"{{language}}\"></div></em></div></div></div></div>"
+    "<div class=\"skos-concept\"><div class=\"top top-alt\"><span ng-if=\"notation.length\" class=\"notation\">{{notation[0]}}</span> <b><span ng-if=\"prefLabel\" skos-label=\"concept.prefLabel\" lang=\"{{language}}\"></span></b> <a ng-if=\"uri && uri != notation\" class=\"uri\" href=\"{{uri}}\" target=\"_blank\"><span class=\"glyphicon glyphicon-globe\"></span></a></div><div ng-if=\"altLabel\" class=\"skos-concept-altlabel\"><ul><li ng-repeat=\"a in altLabel[language]\"><span ng-if=\"$index < 5\">{{a}}</span> <span style=\"margin-left:-4px;margin-right:3px\" ng-if=\"$index < 4 && $index < altLabel[language].length-1\">,</span></li></ul></div><div ng-if=\"broader.length || narrower.length || related.length\" class=\"skos-concept-connected\"><div ng-if=\"broader.length\" class=\"skos-concept-relation skos-concept-relation-broader\"><ul ng-repeat=\"c in broader\"><li><div>&#8598;</div><div class=\"skos-concept-relation-label\"><span ng-if=\"c.prefLabel\" skos-label=\"c.prefLabel\" lang=\"{{language}}\" ng-click=\"click(c)\" title=\"{{c.notation[0]}}\"></span> <span ng-if=\"!c.prefLabel\">{{c.notation[0]}}</span></div></li></ul></div><div ng-if=\"narrower.length\" class=\"skos-concept-relation skos-concept-relation-narrower\"><ul ng-repeat=\"c in narrower\"><li><div>&#8600;</div><div class=\"skos-concept-relation-label\"><span ng-if=\"c.prefLabel\" skos-label=\"c.prefLabel\" lang=\"{{language}}\" ng-click=\"click(c)\" title=\"{{c.notation[0]}}\"></span> <span ng-if=\"!c.prefLabel\">{{c.notation[0]}}</span></div></li></ul></div><div ng-if=\"related.length\" class=\"skos-concept-relation skos-concept-relation-related\"><ul ng-repeat=\"c in related\"><li><div>&#8594;</div><div class=\"skos-concept-relation-label\"><span ng-if=\"c.prefLabel\" skos-label=\"c.prefLabel\" lang=\"{{language}}\" ng-click=\"click(c)\" title=\"{{c.notation[0]}}\"></span> <span ng-if=\"!c.prefLabel\">{{c.notation[0]}}</span></div></li></ul></div></div><div style=\"margin-bottom:10px\"><div skos-notes=\"concept\" lang=\"{{language}}\"></div></div></div>"
   );
 
 
   $templateCache.put('template/skos-list.html',
     "<ul ng-if=\"concepts.length\" class=\"skos-simple-list\"><li ng-repeat=\"c in concepts\"><div class=\"set\" tabindex=\"0\" ng-keydown=\"onKeyDown($event, $first, $last, $index)\" list-index=\"{{listname + '_' + $index}}\" ng-focus=\"onFocus($index)\"><span style=\"whitespace:nowrap\" class=\"notation skos-list-notation\" title=\"{{popOverTitle(c.prefLabel)}}\" ng-click=\"onClick($index)\">{{c.notation[0]}}</span> <span ng-if=\"showLabels\" skos-label=\"c.prefLabel\" lang=\"{{language}}\" class=\"skos-list-label\" ng-click=\"onClick($index)\"></span><div style=\"display:inline-table;padding-left:3px\"><a ng-if=\"onSelect\" href=\"\" ng-click=\"onSelect(c)\" style=\"text-decoration:none;vertical-align:middle\"><span class=\"glyphicon glyphicon-info-sign\" title=\"Select concept (ENTER)\"></span></a> <a ng-if=\"canRemove\" href=\"\" ng-click=\"removeConcept($index)\" style=\"text-decoration:none;vertical-align:middle\"><span class=\"glyphicon glyphicon-trash\" title=\"Remove concept (DEL)\"></span></a></div></div></li></ul>"
+  );
+
+
+  $templateCache.put('template/skos-notes.html',
+    "<ul ng-repeat=\"n in scopeNote\" class=\"skos-note skos-note-scopeNote\"><li skos-label=\"n\" lang=\"{{language}}\"></ul><ul ng-repeat=\"n in definition\" class=\"skos-note skos-note-definition\"><li skos-label=\"n\" lang=\"{{language}}\"></ul><ul ng-repeat=\"n in example\" class=\"skos-note skos-note-example\"><li skos-label=\"n\" lang=\"{{language}}\"></ul><ul ng-repeat=\"n in historyNote\" class=\"skos-note skos-note-historyNote\"><li skos-label=\"n\" lang=\"{{language}}\"></ul>"
   );